Implementing Gates for authorization in Backend

+++ app/Http/Controllers/PostController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Category;
 use App\Models\Post;
 use Illuminate\Http\Request;
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\File;
 
 class PostController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -23,6 +24,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create_post');
         $categories = Category::all();
         return view('create',compact('categories'));
     }
@@ -32,6 +34,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create_post');
+
         $request->validate([
             'image' => ['required', 'max:2028', 'image'],
             'title' => ['required', 'max:255'],
@@ -66,6 +70,7 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('edit_post');
         $post = Post::find($id);
         $categories = Category::all();
         return view('edit',compact('post','categories'));
@@ -76,6 +81,8 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('edit_post');
+
         $request->validate([
             'title' => ['required', 'max:255'],
             'category_id' => ['required', 'integer'],
@@ -107,6 +114,8 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete_post');
+
         $post = Post::findOrFail($id);
         $post->delete();
 
@@ -114,11 +123,15 @@ class PostController extends Controller
     }
 
     public function trashed(){
+        $this->authorize('create_post');
+
         $posts = Post::onlyTrashed()->paginate(10);
         return view('trashed',compact('posts'));
     }
 
     public function restore($id){
+        $this->authorize('create_post');
+
         $post = Post::onlyTrashed()->findOrFail($id);
         $post->restore();
 
@@ -126,6 +139,8 @@ class PostController extends Controller
     }
 
     public function forceDelete($id){
+        $this->authorize('create_post');
+
         $post = Post::onlyTrashed()->findOrFail($id);
         File::delete(public_path($post->image));
         $post->forceDelete();
